backend/requirements.txt
blinker
click
Flask
Flask-Migrate
Flask-Cors
Flask-SQLAlchemy
itsdangerous
Jinja2
MarkupSafe
python-dotenv
SQLAlchemy
typing_extensions
Werkzeug
psycopg2-binary
openai
beautifulsoup4
gunicorn
requests
browser-use
playwright
langchain